Extract GUI manager for desktop app
