+++ src/main/java/laifu/fu/lai/FuDesktopApp.java
@@ -8,10 +8,6 @@ import org.cef.CefApp;
 import org.cef.CefClient;
 import org.cef.browser.CefBrowser;
 
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.WindowAdapter;
-import java.awt.event.WindowEvent;
 import java.io.File;
 import java.io.IOException;
 import java.util.logging.Level;
@@ -49,20 +45,7 @@ public class FuDesktopApp {
         CefApp cefApp = builder.build();
         CefClient client = cefApp.createClient();
         CefBrowser browser = client.createBrowser("http://localhost:8080", false, false);
-        Component browserUI = browser.getUIComponent();
 
-        JFrame frame = new JFrame("Fu Desktop App");
-        frame.getContentPane().add(browserUI, BorderLayout.CENTER);
-        frame.setSize(1024, 768);
-        frame.setLocationRelativeTo(null);
-        frame.addWindowListener(new WindowAdapter() {
-            @Override
-            public void windowClosing(WindowEvent e) {
-                CefApp.getInstance().dispose();
-                frame.dispose();
-                System.exit(0);
-            }
-        });
-        frame.setVisible(true);
+        new GuiManager().showBrowser(browser);
     }
 }
